Moved installation from README to installation.md
The `README.md` now contains only instruction for installation
from the RPM package (the primary way how ordinary user should install
DiffKemp). The instruction for installation from source have been moved
to `installation.md` which is linked from the README.

For the RPM package is added example how to install it by commands.

Changes in the text of `Installation from source` section:
- Updated supported LLVM versions (added 16, 17).
- Added example commands for installing dependencies on Fedora
  and Ubuntu.
- Added note about that the system can provide newer LLVM that we
  do not yet support and it is necessary to install the version which
  we support.
- Shortend information about dependencies for the build system and
  the result viewer,
- Simplified the building instructions and added the flag for building
  the result viewer (added a note that it can be omitted if user does
  not want to use it).
- Added an example how to install DiffKemp to system after building it.

Note: The usage of `pip install ...` should be probably changed
because of PEP 668 [1], which tries to make user use rather virtual
environment than installing packages directly to system because it
can break software shipped with OS.
In Ubuntu 24.04 is already necessary use `--break-system-packages`
flag to be able to use pip.
For now added note to `pip install -r requirements.txt` that the
user may want to use it in virtual environment.

[1]: https://peps.python.org/pep-0668/