pyproject: move to pyproject from setuptools
Python community is moving towards storing metadata for projects inside
TOML files. With this commit we are also moving into this direction but
unfortunately due to cffi we cannot move to `pyproject.toml` as the cffi
developers do not plan to support it currently.

We also can get rid of `requirements.txt` as it is now obsolete, as we
can define the dependencies inside of `pyproject.toml` and they are
installed automatically with `pip install .` (there is no need to use
`-r` option separately).

It is also possible to generate diffkemp executable automatically with
pip.