Viewer: Fix showing empty context line below diff
There was a bug in the `react-diff-view` package which we are
using for rendering the code in the viewer.
If a diff contained an empty context line below the diff, the bug was
causing the line was not rendered. When the rest of the code was
expanded the line was still missing and sometimes the last shown line
was duplicated.

The bug was fixed in 3.2.1 version of the package by [1] PR.
The bug could be encountered e.g. between `4.18.0-193.el8` and
`4.18.0-240.el8` version when viewing code for
`schedule_hrtimeout_range_clock` differing function in
`schedule_hrtimeout_range` compared function.

Version 3.2.0 changed rendering of empty lines by rendering it
as space [2]. This required updating our tests.

[1] https://www.github.com/otakustay/react-diff-view/pull/216
[2] https://www.github.com/otakustay/react-diff-view/issues/210