Viewer: Fix showing empty context line below diff
There was a bug in the `react-diff-view` package which we are
using for rendering the code in the viewer.
The bug caused empy context line below diff to not be rendered.
The bug was fixed in 3.2.1 version of the package by [1] PR.

Version 3.2.0 changed rendering of empty lines by rendering it
as space [2]. This required updating our tests.

[1] https://github.com/otakustay/react-diff-view/pull/216
[2] https://github.com/otakustay/react-diff-view/issues/210