Viewer: Hiding warning about missing source map
After upgrading the `react-diff-view` package to 3.2.1 version,
a warning about failing to parse the source map for the
`react-diff-view` package appears when running the viewer in development
mode and when building the production version.
This commit hides the warning in development mode (when running
`bin/diffkemp view --devel ...`) by setting `GENERATE_SOURCEMAP=false`
preventing the console to be polluted with the warnings.