Fix getPointeeStructTypeInfo for multiple dbg info
We are using `getPointeeStructTypeInfo` function in the `cmpMemset`
and `cmpAllocs` methods when the set/allocated size is not the same
in both compared functions. With the help of this function, we check
that the set/allocated type names are equal and the set/allocated size
corresponds to the size of the type. For LLVM 15 and above (when opaque
pointers became enabled by default), we use debug intrinsics and debug
metadata to get the type to which the pointer in memset/allocs points.

It can happen that there is multiple debug metadata for the same value.
The other ones can have different scope and describe different variable
e.g. parameter of called function which was inlined (e.g. the C `memset`
function). These debug data does not have to contain necessary
information about the pointee type (because they 'cast' it to `void *`).
Currently, we were always taking the first debug metadata (respectively
debug intrinsic), this could cause that we were trying to extract the
pointee type from metadata which does not contain it. This could cause
assert/abort caused by trying to `dyn_cast` a `nullptr`, this was fixed
in previous commit. Currently it causes a false positive result.

An example causing this issue could look like this:

```c
// old
struct s {
    int a;
    int b;
};
void f(struct s *var, char x) {
    memset(var, x, sizeof(struct s));
}
// new
struct s {
    int a;
    int b;
    int c;
};
void f(struct s *var, char x) {
    memset(var, x, sizeof(struct s));
}
```

The issue can be replicated when using the
`nix develop .#diffkemp-llvm17` environment and the programs
are built using
`diffkemp build --clang-append="-O2" --clang-append="-g" --no-opt-override`
command and then compared.

Snippet of the resulting LLVM IR for the old version looks like this:

```llvm
define void @f(ptr %0, i8 %1) {
  ; ...
  ; multiple dbg value intrinsic for %0 destination value of memset intrinsic
  call void @llvm.dbg.value(metadata ptr %0, metadata !667, metadata !DIExpression())
  call void @llvm.dbg.value(metadata ptr %0, metadata !670, metadata !DIExpression())
  tail call void @llvm.memset.p0.i64(ptr %0, i8 %1, i64 8, i1 false)
  ; ...
}
; dbg metadata for %0 value with scope of "f" function, contain info about
; the pointee type
!667 = !DILocalVariable(name: "var", arg: 1, scope: !656, type: !659)
!656 = distinct !DISubprogram(name: "f", ...)
!659 = !DIDerivedType(tag: DW_TAG_pointer_type, baseType: !660, size: 64)
!660 = distinct !DICompositeType(tag: DW_TAG_structure_type, name: "s", elements: !661)
!661 = !{!662, !664}
!662 = !DIDerivedType(tag: DW_TAG_member, name: "a", baseType: !663, size: 32)
!663 = !DIBasicType(name: "int", size: 32, encoding: DW_ATE_signed)
!664 = !DIDerivedType(tag: DW_TAG_member, name: "b", baseType: !663, size: 32)
; dbg metadata for %0 value, inlined from memset funtion (scope),
; does not contain info about pointee type
!670 = !DILocalVariable(name: "__dest", arg: 1, scope: !671, type: !674)
!671 = !DISubprogram(name: "memset", ...)
!674 = !DIDerivedType(tag: DW_TAG_pointer_type, baseType: null, size: 64)
```

This commit tries to fix it by using debug intrinsic which has
the same scope as is a current function name instead of using the
first debug intrinsic. It is unknown if this solution is universal
for all cases but in worst case it should hopefully give non-equal
result and not abort the program.

In unit tests were updated `cmpMemset` and `cmpAllocs` test cases
to include the `DISubprogram` information so the test works correctly.
Also test for the multiple debug info was added.

The commit also updates version of LLVM that is used for running
SimpLL unit tests with address sanitizer enabled to LLVM 17.
The reason is that `DIBuilder::createFunction` method which was added
to the unit tests contained memory leak until version LLVM 17 [1].

[1] https://www.github.com/llvm/llvm-project/commit/9eda46b1817f60378143846477ef71f0131a121d