CI: Check DiffKemp builds by running their binary
We allow multiple ways how to use DiffKemp:
- Using DiffKemp in Nix development environment,
- Using it locally
  - for development purposes (using `pip install -e`)
  - as 'release' (using `ninja install && install bin/diffkemp ...`)
This commits adds 'all' possible way how to build/use DiffKemp
to CI, where it checks that the build is working by using the build
binary for bulding and comparing make-based project which is parts of
tests directory.

Note: For local builds is used ubuntu-22.04 GitHub runner instead of
ubuntu-latest because GitHub will be soon using Ubuntu 24.04 as
ubuntu-latest. Ubuntu 24.04 follows PEP-668 [1] where it is harder
to use `pip install` command - it is necessary to use option
`--break-system-packages` or use virtual environments, ...

[1] https://peps.python.org/pep-0668/