Encode LLVM instructions into SMT problem
We encode the semantic equivalence of the detected blocks into an SMT
formula as follows. The final formula has the form precondition &&
block1 && block2 && !postcondition, where precondition encodes the
equivalence of input variables of the block (based on varmap), block1
and block2 encode the semantics of the compared blocks and postcondition
encodes the equivalence of output variables of the blocks (i.e. those
variables that are used outside the block).

This commit adds implementation of the precondition and encoding of
blocks. The interesting part is encoding of blocks, where we make use of
LLVM IR's SSA property to build a new SMT variable for each LLVM IR register.
The encoding makes use of pointer addresses of the LLVM instructions.
Using this encoding, we can conveniently compare the operation performed
by each instruction type.