Add support for LLVM 19
This commit adds a support for LLVM 19, the main differences in
comparison to the previous version are:

1. The removal of `equals` method from `StringRef` class.
2. The removal of `getPredicate` from `ConstExpr` class.
3. Various changes in includes, that require us to use more fine-grained
   include strategy.
4. The separation of some analysis functions and classes into a separate
   header.
5. The deprecation of debug intrinsics has begun.

The `equals` method was removed in this commit [1] and the equality
operator is in the code base from LLVM 3, so it should be safe to use.

The name of some passes has changed in [2]. Therefore, we have to check
for some of these passes and change their name in accordance to the used
LLVM version.

Because of that we have to check the version with which was the DiffKemp
SimpLL library built. This required addition of new function to the FFI.

The `getPredicate` method was removed from `ConstExpr` class in commit [3],
we have extracted the method for getting an instruction from expression
into a stand-alone function, but this method has changed upstream.
Therefore, I have extracted the new changed method into a new function.

The analysis classes and functions were moved into a new header in [4].

The LLVM aims to remove debug intrinsics, as explained in [5]. However,
we still rely on them in our tests. Therefore, we have to set the debug
type back to the old one as our code was affected by [6].

[1]: https://github.com/llvm/llvm-project/commit/3fa409f2318ef790cc44836afe9a72830715ad84
[2]: https://github.com/llvm/llvm-project/commit/e390c229a438ed1eb3396df8fbeeda89c49474e6
[3]: https://github.com/llvm/llvm-project/commit/deab451e7a7f2dff42097049274637052c87eabd
[4]: https://github.com/llvm/llvm-project/commit/a4cd9812bacf959ba7a1e7dbb022b7546f5952e3
[5]: https://llvm.org/docs/RemoveDIsDebugInfo.html
[6]: https://github.com/llvm/llvm-project/commit/9997e0397156ff7e01aecbd17bdeb7bfe5fb15b0